Split graph and table routes in to separate files
