+++ server/routers/index.js
@@ -3,5 +3,7 @@ module.exports.api = {};
 module.exports.api.query	= require('./api/query');
 module.exports.api.export   = require('./api/export');
 
-module.exports.dashboard    = require('./dashboard');
+module.exports.table		= require('./table');
+module.exports.graph		= require('./graph');
+
 module.exports.optInOut     = require('./opt-in-out');
